Pipeline for validating and planning Terraform code
This commit introduces code that is meant to validate
(i.e. `terraform fmt` and `terraform validate`) and ultimately run
a plan on Terraform code within a known implementation directory. This
pipeline is meant to be used with a Monorepo of code that has many
different individual Terraform implementation directories.

To track whether a directory within the repository is a known
Terraform implementation directory, a file called
terraform-directory-mappings.yaml has been introduced to keep track of
all of those directories. This file will also allow for metadata around
each of those implementation directories, including team ownership and
anything else that would be useful later on.

When a PR is opened, the changed files within the PR are scanned to
determine their parent directories. Those parent directories are
then compared with the known Terraform implementation directories, and
any matches will trigger steps to perform Terraform validation and,
pending successful validation, a `terraform plan` being run against that
directory. The output of the `terraform plan` will be returned as
a comment to the PR so that reviewers can see what is meant to happen.
Automation of `terraform apply` is not currently being done - that will
continue to be done by hand until that level of automation is required.