Restructure and document the pipeline
This commit removes the explicit node block in favor of using the agent
block for determining the executor that will need to run the code.
An explicit git checkout stage has been added to ensure that the code is
checked out on the executor before anything happens. I'm unsure if the
checkout in the pipeline ONLY runs on the master or if it will be
triggered on the build executor too (if it runs on the executor then we
can strip this stage later - will need more testing). Finally
documentation has been added to the pipeline to make it easier to follow
and understand. We're dealing with a case of wanting to make sure a team
that's new to CI can follow what is going on, so documentation is
important.