Update terraform mappings workflow
The structure and location of the terraform mappings file has changed,
and this commit tracks those changes both in code as well as within the
repository. We're using a list instead of a map for the file's contents
because all we need right now are the paths to the implementation
directories.